+++ crates/client/src/client/datamodels.rs
@@ -16,8 +16,8 @@ impl SplunkClient {
     /// List all data models.
     pub async fn list_datamodels(
         &self,
-        count: Option<u64>,
-        offset: Option<u64>,
+        count: Option<usize>,
+        offset: Option<usize>,
     ) -> Result<Vec<DataModel>> {
         crate::retry_call!(
             self,
