+++ crates/client/src/endpoints/search/saved.rs
@@ -27,8 +27,8 @@ pub async fn list_saved_searches(
     client: &Client,
     base_url: &str,
     auth_token: &str,
-    count: Option<u64>,
-    offset: Option<u64>,
+    count: Option<usize>,
+    offset: Option<usize>,
     max_retries: usize,
     metrics: Option<&MetricsCollector>,
 ) -> Result<Vec<crate::models::SavedSearch>> {
