+++ crates/client/src/endpoints/workload.rs
@@ -48,8 +48,8 @@ pub async fn list_workload_pools(
     client: &Client,
     base_url: &str,
     auth_token: &str,
-    count: Option<u64>,
-    offset: Option<u64>,
+    count: Option<usize>,
+    offset: Option<usize>,
     max_retries: usize,
     metrics: Option<&MetricsCollector>,
 ) -> Result<Vec<WorkloadPool>> {
@@ -113,8 +113,8 @@ pub async fn list_workload_rules(
     client: &Client,
     base_url: &str,
     auth_token: &str,
-    count: Option<u64>,
-    offset: Option<u64>,
+    count: Option<usize>,
+    offset: Option<usize>,
     max_retries: usize,
     metrics: Option<&MetricsCollector>,
 ) -> Result<Vec<WorkloadRule>> {
